requirements.txt
scipy
torchvision
mkl
mkl-fft
intel-openmp
numpy
ujson
git+https://github.com/pytorch/tnt.git@master
torchfold